### 2024-09-05 #654 @WeichengSun
ILSM: Analyze Interconnection Structure of Multilayer Interaction Networks
Submitting Author Name:  SunWeicheng
Submitting Author Github Handle: @WeichengSun
Repository:  https://github.com/WeichengSun/ILSM
Version submitted: 1.0.3.2
Submission type: Stats

Version accepted: TBD

---

```
Package: ILSM
Title: Analyze Interconnection Structure of Multilayer Interaction Networks
Version: 1.0.3.2
    person("WeiCheng", "Sun", , "sunwch2023@lzu.edu.cn", role = c("aut", "cre")),
    person("Chuan", "Yan", , "yanchuan@lzu.edu.cn", role = c("aut")),
    person("Yangyang", "Zhao", , "120220907711@lzu.edu.cn", role = c("aut")))
    In view of the analysis of the structural characteristics of the multilayer network has been complete, however, there is still a lack of a unified operation that can quickly obtain the corresponding characteristics of the multilayer network.
    To solve this insufficiency, 'ILSM' was designed for supporting calculating such metrics of multilayer networks by functions of this R package.
    Matrix,
    igraph

```

- Data Packages
- Statistical Packages
     - [x] ecological network analysis
     - [ ] spatial analysis
     - [ ] time series analysis

This package was designed for supporting calculating such metrics of multilayer networks.

Reseachers who work for ecological interaction effects.

Yes!
