### 2024-09-05 #654 @adamhsparks
Hi @WeichengSun, it seems that the bot is having some issues, I'll see what we can do about that. Do you still wish to proceed with the review?

If so, you'll need to set up [CI](https://ropensci.github.io/dev_guide/ci.html) with test coverage reporting and ensure all of your exported functions have examples in your [ROxygen code](https://ropensci.github.io/dev_guide/building.html#documentation) as you've not indicated that these criteria are met.
