### 2024-09-06 c01deeb
Delete Contributing.md
