+++ README.Rmd
@@ -15,6 +15,7 @@ knitr::opts_chunk$set(
 
 <!-- badges: start -->
 
+[![R-CMD-check](https://github.com/WeichengSun/ILSM/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/WeichengSun/ILSM/actions/workflows/R-CMD-check.yaml)
 <!-- badges: end -->
 
 ILSM is designed for analyzing interconnection structures including interconnection patterns, versatility and motifs in multilayer interaction networks.
