+++ pkgdown.yml
@@ -1,0 +1,6 @@
+pandoc: 3.1.11
+pkgdown: 2.1.0
+pkgdown_sha: ~
+articles:
+  motif_guideline: motif_guideline.html
+last_built: 2024-09-06T04:05Z
