+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.1.0
 pkgdown_sha: ~
 articles:
   motif_guideline: motif_guideline.html
-last_built: 2024-09-06T04:05Z
+last_built: 2024-09-06T04:07Z
