### 2024-09-06 #654 @adamhsparks
@WeichengSun, the package checks indicate that the test coverage needs to be improved before we can start the review but everything else checks out.

Would you like to proceed? If so, just ping me in this Issue when you've got the package's test coverage up to 75% and I'll take it from there. Also, feel free to ask me if you have any other questions.
