### 2024-09-07 3ef3a1b
Update coid.R
