### 2024-09-07 2e9cf04
Create test-coid.R
