### 2024-09-07 d2ea121
Create test-cois.R
