### 2024-09-07 d4f8ce6
Update test-igraph_from_matrices.R
