### 2024-09-07 8e67eef
Create test-null_model.R
