### 2024-09-07 0af56b1
Create test-poc.R
