### 2024-09-08 #654 @adamhsparks
You have it on CRAN, already, so it shows unavailable.
