### 2024-09-08 #654 @adamhsparks
@WeichengSun, could you please set up or fix the test coverage for GitHub Actions and add a badge for it? This indicates that it's not passing, https://github.com/WeichengSun/ILSM/actions/runs/10752142570
