+++ .github/workflows/test-coverage.yaml
@@ -59,3 +59,9 @@ jobs:
         with:
           name: coverage-test-failures
           path: ${{ runner.temp }}/package
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v4
+        with:
+          file: ./coverage.xml
+          token: ${{ secrets.CODECOV_TOKEN }}
+
