+++ NAMESPACE
@@ -13,7 +13,6 @@ export(null_model)
 export(pc)
 export(poc)
 export(role_sim)
-export(subnet_cor)
 import(igraph)
 importFrom(Matrix,Diagonal)
 importFrom(Matrix,Matrix)
