### 2024-09-12 #654 @WeichengSun
Hi @adamhsparks, I want to submit an Applications Article about my package to [Methods in Ecology and Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/), but I don't know how my package can pass the review of ropensci. Please give me some suggestions. Thank you.
