### 2024-09-14 #654 @adamhsparks
Hi @WeichengSun, after checking with the editorial board, I was alerted that you seemingly have mashed together our pre-submission and regular stats templates with your submission.

Could I ask you to please close this and open a regular stats submission so that the bot picks up this as a stats submission?

You will need to add our stats standard annotations to the package as well. The bot will pick that up and the process won't proceed until you've done that, so I suggest doing that before opening a new issue and using the proper stats template.
