### 2024-09-14 #654 @WeichengSun
Hi @adamhsparks, what can I refer to for 'stats standard annotations' you mentioned? I am very confused about this.
