### 2024-09-14 #654 @adamhsparks
Hi @WeichengSun, I'd start with our [rOpenSci Statistical Software Peer Review](https://stats-devguide.ropensci.org/) guide, specifically [Chapter3](https://stats-devguide.ropensci.org/pkgdev.html#general-standards) for these details.
